ZMQ: Improved reliablity of the hashblock notification
Made it so that if the notification arrives while we were already in the
synch task, we remember the hash and then check it later against our
concept of latest tip. If it doesn't match, we immediately start the
synch task again right away.

This was we can get the latest blocks always (or be notified of reorgs!),
even if we were busy synching mempool when the zmq notification arrived.