Fixups to Util::AsyncSignalSafe::SBuf
Cleaned up the code for append(Integral) to branch less, and also truncate
rather than refuse to print if not enough space (this is closer to what client
code would expect). It uses a temporary buffer on the stack for
processing but due to cache and locality of reference it turns out that
it's fast anyway (before we were doing copying anyway to reverse it).

Overall it runs a slight bit faster than before.