Added a static_assert: maxHistoryMax is limited by 2GiB QByteArray limit
This static_assert ensures that if we ever change Options::maxHistoryMax
to a larger value, that we get an error from the compiler since our code
would potentially fail at runtime.  Better to error-out at compile-time
than at runtime!

The assert is needed because it's not safe to construct QByteArrays of size
beyond 2GiB.

This limitation corresponds to a limit of about ~27M to maxHistoryMax.
(We use 25M for this limit now).

This static_assert was added in the interests of defensive programming.