Add support for bitcoind ".cookie" file authentication
Fulcrum now supports bitcoind's auto-generated ".cookie" file,
which is created each time bitcoind is restarted if no rpcuser= and
rpcpassword= exists in bitcoin's .conf file.

To use cookie-file-based auth instead of a hard-coded user/pass, simply
specify --rpccookie or -K on the CLI and supply the full path of the
".cookie" file that bitcoind generates.  This file normally lives in the
datadir/.cookie.