docker: Added support for building Windows static binary with Qt6
This is for the docker build mechanism (for releases).  A new "platform"
was added, "win_qt6". This will use a different docker file to build the
binary using Qt6 6.2.0-beta3, rather than Qt 5.15.2.

Unfortunately, the produced binary no longer works on Windows 7.  It was
tested on Windows 10 and appears to work ok there though. It may be the
case that Qt6 simply stopped supporting Win7.  C'est la vie.