Fix spurious Error message in ServerBase::attachPerIPDataAndCheckLimits
It turns out that the low-level socket may reach the unconnected state
before even Qt processes this socket event. In that case, calling
socket->peerAddress() will return a null address -- since the kernel
thinks the socket is no longer connected, even though Qt thinks
the socket is still connected.

In the rare even that this corner-case would occur, we were spaming an
"Error" to the log (only happens with misbehaving clients or port
scanners that just connect and then immediately disconnect).  We have
fixed the situation here to just print a "Warning" level message when
this corner-case is reached, rather than an ugly RED "Error" message.