Workaround to support Phoenix wallet
It turns out the Phoenix BTC wallet sends slightly out-of-spec
JSON-RPC messages. They include the key "jsonrpc": "1.0" which is
completely out-of-spec.  Fulcrum was not liking this at all and
rejecting such clients.  However, since Phoenix can't be changed, and
since it has users, we must be tolerant here and just accept their
messages and not panic.

This fix closes #91.