Fixed banner file variable substitutions to not recurse
There was a corner case quasi-bug in the way variable substitutions
worked (e.g. $SERVER_VERSION), in that if the substituted text itself
container a variable "$SERVER_SUBVERSION", etc, it would recurse and
replace that as well.  This is unexpected, and has been fixed. Also, the
code that does the variable substitutions has been made more performant.

Additionally, a test has been added to test the banner file substitution
mechanism for regressions

Also in this commit: Various small nits