Add JSON-RPC batching + increase some resource defaults + bump version to 1.6.0 (#97)
* Added Json::estimateMemoryFootprint

This maybe will be used to limit batch requests (and responses) to a certain
maximum complexity.

* Made _send* functions in RPC::ConnectionBase no longer virtual

Nothing in the codebase was overriding them, so they are best left
non-virtual. In the original design of the code I intended for
subclasses to customize these functions, but the design of the system
went in a direction where lots of behavior is controlled by
RPC::ConnectionBase anyway.

Making them non-virtual should hopefully shave a few cycles when
processing messages.

* Misc. fixups

- Got rid of ProcessObjectResult::Error:disconnect field (can be deduced
  from conn.errorPolicy)
- Bug, fixed counts: ++conn.nRequestsSent -> ++conn.nResultsSent
- Added some commments
- Only call AGAIN() if conn.isGood()

* Added stats printing for BatchProcessors

Also some nits, renamed "submittedRequests" to "unansweredRequests", and
also don't sent result to client unless conn.isGood().

* Added bitcoind request throttling support for batch requests

The batch request just pauses itself whenever isReadPaused() is set to
true.

* added a comment

* Added "max_batch", per-ip limit on the number of extant batch requests

This is a DoS measure. The current default limit is 200 (maybe tune
this?).

Currently the code has too many debug prints (these will be removed
before merge to master).

* nitted a comment

* Fixups: Slight fix in BatchProcessor::process() to ensure correctness

It's possible that calling `emit gotMessage()` etc will lead to the
`batch.responses` array being appended immediately.  The algorithm we had
assumed it wouldn't change from underneath out feet, when it can.

This commit fixes that.

* Fixed some typos

* added a comment

* Added `options` reference to class Client

Also in this commit: Reorganized the order of some struct member
variables to pack the Client and ConnectionBase classes tighter and
have them waste less memory.

* Set default batch limit to 345 (seems safe for BlueWallet at least)

* Suppress a Qt6 compile warning

* Refactor to allow for killing a BatchProcessor job if it exceeds limits

Jobs can now be killed if they exceed their limits while they are
preparing results.  The client will be sent an error message for the
entire batch, and the job immediately stops processing.

* Added batch request memory cost control and tracking

Just to prevent memory DoS or other nonsense, the batch requests cannot
exceed `max_buffer` in terms of the json data size for the both the
batch request itself + the results.

This tracking is done on a per-IP basis.  Batch requests that exceed the
`max_buffer` limit will be killed immediately.

As a batch request is processed, and as results accumulate, the memory cost
is updated to reflect the accumulated json data result byte size.

The memory costs are rough estimates, and are not exact, but they are good
enough approximations for the purposes of guarding against DoS.

* Disallow `null` "id" in JSON-RPC batch requests

Due to our architecture, they can be ambiguous and can clash with error
responses coming through.  So we just disallow them from batches.

Hopefully no well-written client will be sending `null` "id" for batch
requests anyway, so I don't anticipate this limitation being a problem.

* Added "batchZombies" mechanism to filter responses for dead batches

This commit adds a filter for asynchronous results for "batch zombies".
That is, messages that were emitted via gotMessage() but for which the
batch processor was killed before a response was generated asynchronously.
We need to filter these out from the client, to avoid a buggy corner case.

* A nit in batch request timeout code

* Got rid of some TODOs

* Tweak

* Added documentation for `max_batch`

Added to fulcrum-example-config.conf.

* Nit in Servers.cpp (always take batch name)

This is because we actually may warn with it.

* A tweak - update lastGood only if a real message arrived

This is to make the new code behave identical to old code before the
batching feature addition.

* Some cleanup

- Compile-out excessive debug prints (only enabled with a constexpr bool)
- Deleted some TODOs that don't need to be there

* Bumped version to 1.6.0

This is on account of the new JSON-RPC batching support which warrants a
new minor version number.

* Increased defaults by for some config options

- db_max_open_files default is now 40 (was 20)
- db_mem default is now 512 (was 420)
- max_buffer is now 8000000 (was 4000000)
- workqueue is now 15000 (was 10000)
- max_subs_per_ip default is now 75,000 (was 50,000)

* Controller: some tweaks to block prefetch in the BTC case

We can prefetch more on BTC than on BCH since it has smaller blocks.