Batching: Fix for issue when compiling with GCC
It turns out GCC interprets the expression items{someItems} if both
items and someItems are QVariantList as a std::initializer_list
constructor rather than as a copy or move c'tor. Ugh.

This was causing bugs for the batch processor (was unusable) on GCC.

On clang this works perfectly. I had been using clang as the compiler,
so I never detected the problem.

Closes issue #99