Fix for ZMQ notifier to also understand "*" URL notation
Basically sometimes the remote's ZMQ URL may look like:
`tcp://*:8330` (when it is binding to "any" address).
We have to rewrite this "*" to the known IP of the remote bitcoind (as
we already do for e.g. 0.0.0.0).

This commit fixes an issue seen by a user; see issue #102 for the full
discussion.