Blocks DL: Improve reliability of bitcoind connection on slow systems or with huge blocks + misc fixes (#120)
This commit should fix issue #116. It also addresses the Bitcoin Verde Scaling Report 
dated May 2022 which revealed that Fulcrum has some issues downloading very large
blocks (256MB+) in a series. The two issues are related, and are basically because the
following could happen on a slow HDD system or with a bitcoind that is serving up 
very large blocks: