Force static linking of openssl lib for linux static builds
We ensure that Qt 5.15.6's QtNetworking statically links to libssl.a and
libcrypto.a.  This fixes a bunch of issues when running the released
static binary on newer systems.  Closes #126.

This changs affects only the "linux" and "newlinux" docker build targets,
and not the older "oldlin" target, which still relies on system openssl.