Nuke the "zero after free" allocator from our bitcoin/ sources
It's useless and just wastes cycles. The bitcoin "secure" allocator
would be appropriate if we handled private keys, but we do not. The
simple "zero after free" allocator is just a performance eater.

It took me this long to notice it exited. It's gone. Class
`bitcoind::CDataStream` now just uses a `std::vector<char>` rather than
a custom vector with a custom allocator.