Got rid of compiling with -fomit-frame-pointer
It's asking for compiler bugs to crop up. (It caused problems on my
macOS clang).  Also not needed on newer CPUs since they are not
register-impaired such as i686 was.

Also in this commit: a nit, call the specific std::string c'tor of
superclass std::runtime_error directly and don't rely on overload
resolution for it.