Fixed very rare bug where BitcoinD instances may become unresponsive
The symptom is one of the 3 BitcoinD instances in the app would stop
working and continually attempt to reconnect to the remote bitcoind
daemon, never quite succeeding and never receiving a ping reply.

The way it could be triggered is if the remote daemon drops conn on us
at an inopportune time while we are reading content from its HTTP
response. The internal StateMachine used by this class would end up in a
confused state.

The fix is to implement `on_disconnected` in the HttpConnection class and
be sure to reset the StateMachine for this instance on a connection
drop. This fixes the bug nicely.

The bug was extremely rare but not impossible.