Refactor ByteView class & refactor bitcoin tests
ByteView class was inheriting from std::string_view which can cause
problems since its char_type is std::byte (which is not officially
standards-compliant to do). This fixes compile warnings on clang-18.

Also, in this commit we refactor the bitcoin tests and create a
subdirectory called tests/ which has a header "Tests.h" which allows us
to do BCHN-like BOOST_CHECK() and other macros... it emulates the BOOST
test framework, essentially, in a hacky but working way.