Introduce "thread internal name" mechanism for logger, and use it
The existing way we were logging thread names was to query the
QThread::currentThread() object name. But this is not thread safe
because the QThread::currenntThread() object may not be "living in" the
current thread! As such, this could lead to crashes in some cases (in
particular with the CoTask).

Instead, we introduce the concept of a "thread internal name" with
getters/setters in the Util::ThreadName namespace.

All ThreadObjectMixins and other thread objects in the app explicitly
must set their thread name when they are running, and the logger picks
up the name.

The thread names live in TLS and are thread-safe.