Be explicit about RPA CoTask lifecycle; don't use thread_local
Use of thread_local CoTask to parallelize RPA indexing was added
by a recent commit: e4b01d7dfcfd63135b28933f61757a387b53b419

The situation is fragile and there is potential for UB here due to
static initialization hell.

Instead, we explicitly create the CoTask in the DownloadBlocksTask
thread and it lives as a DownloadBlocksTask member variable, which
is more correct and less bugprone.