ZMQ: Add optional support for "pubhashtx" ZMQ message from bitcoind (#262)
* Added zmqhashtx support for evaluation. Very spammy. May not use it.

* Disabled the hashtx notifier (since it is spammy), plus some fixups

* More tweaks and refactoring to Controller ZMQ mechanism

* Fix hashtx and hashblock handling to enqueue, not call immediately

Hashtx is still way too spammy. It should really be disabled.

* Disabled hashtx again

* Small fixups to restore behavior of hashblock to previous code before these changes

- Call process() immediately if we see any hashblock zmq msg (this was
  what we always did).
- Clear "lask known address" for zmq endpoints that went away
- Misc refactoring

* Made ZMQ hashtx topic default to off, but enableable in config.

This allows users interested in this feature to get it (must be
explicitly enabled).. otherwise default behavior is as before with the
hashtx message not used by Fulcrum (too spammy).

* Don't track "recently seen txids" if zmq hashtx notifs are not enabled

In this way, we save cycles and don't maintain this map if we don't need
the info it provides.