Small nit: Controller::zmqTopicStop no longer needs to be a "slot"
It is no longer being used as a direct slot, but is rather being called
from lambdas and other code. So in the interests of readability we move
this function's reclaration to a different section in Controller.h so
it's no longer a `private slot`.. and now it lives with its sibling
related functions in the same section of the header file.