Increased default db_mem to 1GiB from 512MiB (if unspecified)
It's time. Several years have passed since the original default of
512MiB was determined, and hardware now is even better and it should not
be uncommon to have a 2GiB or above machine. It has been observed that
more db_mem means better performance overall, so we want to be greedy
here. 1 GiB should be a good default for most installs.

Note that we try to not be too greedy and not take up more db_mem than
25% of what the machine reports as having as total physical RAM. We also
don't default to anything less than 512MiB (the older Fulcrum default).