Added a mechanism to detect unclean shutdown conditions
This mechanism depends on the use of a 'latch' file in the datadir
which is created on startup and deleted on clean shutdown. If the file
is found to already exist when it is about to created, a flag is set and
storage uses this flag to know to take some corrective action.

Storage undoes latest block if the unclean shutdown flag is set, as a safety
measure, even though rocksdb guarantees should make this superfluous. Just
in case, it would be safer to undo the latest block in such a situation, if
such a thing is possible, for belt-and-suspenders.

Also in this commit:

- Added a new log type; Alert(), which has bright magenta text and is
  slightly above a Warning but below an Error
- Various small nits and fixups