Redid the readRandomRecords function in DBRecordArray class, plus other small fixes and changes
The previous implementation was confusing and hard to understand, IMHO,
and needlessly complex. I also think this new approach should be faster
in practice due to collating the calls to the DB better and due to use
of MultiGet.

Also in this commit: Added ability for bench/test threads to use the
Log(), and other small fixes and nits.