Redid broadcast_package to do some work in an asynch thread
We validate args in an asynch thread since that is potentially costly
(we must hash the first txn, etc).

When that's done, we forward request to bitcoind.

This required us allowing generic_do_async to accept an optional
"completion" function. In order to make things cleaner and faster we
templatized ServerBase::generic_do_async, and redid its internals
slightly. Hopefully this is cleaner and also hopefully redundant copies
of e.g. std::function are avoided with this template implementation.

Also in this commit: Various nits and fixups in Util.h, added
ThrowInternalErrorIf() macro, and other misc. cleanup.