PeerMgr: Tweaks to TCP fail -> SSL retry processing
Only do the SSL retry if it's clear we failed to connect to the port
either via "Connection refused" or via "timed out" (either of which can
happen if the port is closed, depending on router config on the remote
end).

Also: Tweaked the log message and also clear the preferSsl flag on full-on
failure to retry from scratch with TCP again in future.